+++ Sources/Pouch/Pouch.swift
@@ -1,13 +1,14 @@
 import ArgumentParser
 import Foundation
+import PouchFramework
 
-public struct Pouch: ParsableCommand {
+@main public struct Pouch: AsyncParsableCommand {
     public static var configuration = CommandConfiguration(
         abstract: "A utility tool for secret management",
-        version: "0.2.0",
+        version: "0.4.0",
         subcommands: [Retrieve.self],
         defaultSubcommand: Retrieve.self
     )
-    
+
     public init() {}
 }