+++ Sources/Pouch/main.swift
@@ -1,7 +1,0 @@
-import ArgumentParser
-import Foundation
-import PouchFramework
-
-let logger = Logger(output: .print)
-
-Pouch.main()