+++ Sources/PouchFramework/Models/Cipher.swift
@@ -1,3 +1,4 @@
 public enum Cipher: String, Codable, Equatable {
     case xor
+    case none
 }