+++ Sources/PouchFramework/VariableFetchers/VariableFetching.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public protocol VariableFetching {
-    func fetch(secrets: [SecretDeclaration], completion: (Result<[Secret], Error>) -> Void)
-}